backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Денис Матвеев <alttert@mail.ru>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.41" }
alembic = "^1.16.4"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
yookassa = "^2.4.0"
boto3 = "^1.34.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
httpx = "^0.27.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.10.1"
uvicorn = "^0.35.0"
email-validator = "^2.0.0"
cryptography = "^40.0.2"
pydantic = { extras = ["email"], version = "^2.5.3" }
python-multipart = "^0.0.20"
prometheus-client = "^0.22.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
factory-boy = "^3.3.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
